+++ src/card/cards/control_duration.rs
@@ -4,3 +4,5 @@ use crate::CardDefinitionId;
 
 pub const ALADDIN: CardDefinitionId = CardDefinitionId(1472);
 pub const THRULL_CHAMPION: CardDefinitionId = CardDefinitionId(1473);
+pub const WILLOW_SATYR: CardDefinitionId = CardDefinitionId(1474);
+pub const RUBINIA_SOULSINGER: CardDefinitionId = CardDefinitionId(1475);