+++ src/game/entry_replacements.rs
@@ -469,6 +469,7 @@ impl Game {
             ),
             AppliedEffectDef::CannotBeCountered
             | AppliedEffectDef::DoesNotUntapDuringUntapStep
+            | AppliedEffectDef::MayChooseNotToUntap
             | AppliedEffectDef::CannotBeEnchanted
             | AppliedEffectDef::CannotBecomeEnchanted
             | AppliedEffectDef::CannotChangeController